Move ViewModel related classes from 'module' to 'viewmodel' package
